disabled two field from product create and update
Removed include feature and file include from digital product create and update.

Tag was not working and that is why form was not submitting, fixed also that.

There was a form field that is hidden for digital product creation but had required attribute, for that form was not submitting, fixed that.